% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_one_survey.R
\name{check_and_generate_folder}
\alias{check_and_generate_folder}
\title{Function to check if the folder exist and if not create it}
\usage{
check_and_generate_folder(name)
}
\description{
Function to check if the folder exist and if not create it
}
