% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_one_survey.R
\name{create_issue_folder}
\alias{create_issue_folder}
\title{Function to create a issue folder in the output folder}
\usage{
create_issue_folder(output_folder)
}
\arguments{
\item{output_folder}{Folder where the results need to be saved}
}
\description{
Function to create a issue folder in the output folder
}
