% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_infos.R
\name{extract_location_smart}
\alias{extract_location_smart}
\title{Extract file name information and where the location could be in the smart content}
\usage{
extract_location_smart(name_file, smart_content)
}
\arguments{
\item{name_file}{name of the file}

\item{smart_content}{content of the SMART}
}
\value{
list of information
}
\description{
Extract file name information and where the location could be in the smart content
}
