% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admins_list.R
\name{f_add_element_to_admins}
\alias{f_add_element_to_admins}
\title{Function to add a name in the list of options to match the SMART Survey list
Save the output in the right rda}
\usage{
f_add_element_to_admins(country, list_to_add)
}
\arguments{
\item{country}{Specified country of interest}

\item{list_to_add}{contains eventual name, right_name and level}
}
\description{
Function to add a name in the list of options to match the SMART Survey list
Save the output in the right rda
}
\details{
Add something to the list
}
