% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_with_issue_data.R
\name{f_deal_issues}
\alias{f_deal_issues}
\title{Function to check the different issue files}
\usage{
f_deal_issues(
  country,
  file,
  admin_name,
  log_file,
  smart_folder,
  output_dir,
  country_data
)
}
\arguments{
\item{log_file}{log file}

\item{output_folder}{output_folder}
}
\description{
Function to check the different issue files
}
