% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_infos.R
\name{f_extract_info_smart}
\alias{f_extract_info_smart}
\title{Function to extract the different component of a SMART Survey}
\usage{
f_extract_info_smart(name_file, smart_content, admins_list)
}
\arguments{
\item{name_file}{name of SMART Survey}

\item{smart_content}{SMART Survey content}

\item{admins_list}{obtained through the function f_location_possibility}
}
\value{
[list] list containing: name_admin, level_admin, day/month/year min and max, type_survey.
}
\description{
Function to extract the different component of a SMART Survey
}
\details{
Function to extract information from a unique SMART Survey
}
