% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_content.R
\name{f_extract_metadata_data}
\alias{f_extract_metadata_data}
\title{Function to extraction metadata content of a SMART Survey}
\usage{
f_extract_metadata_data(smart_content)
}
\arguments{
\item{smart_content}{Content of smart survey}
}
\value{
[data.frame] containing the metadata information in the SMART
}
\description{
Function to extraction metadata content of a SMART Survey
}
