% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_content.R
\name{f_extract_mortality_data}
\alias{f_extract_mortality_data}
\title{Function to extraction mortality content of a SMART Survey}
\usage{
f_extract_mortality_data(smart_content, type_survey)
}
\arguments{
\item{smart_content}{Content of smart survey}
}
\value{
[data.frame] containing the mortality information in the SMART
}
\description{
Function to extraction mortality content of a SMART Survey
}
