% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_content.R
\name{f_extract_recall_day}
\alias{f_extract_recall_day}
\title{Function to extraction recall day of a SMART Survey}
\usage{
f_extract_recall_day(smart_content)
}
\arguments{
\item{smart_content}{Content of smart survey}
}
\value{
[data.frame] containing the recall day in the SMART
}
\description{
Function to extraction recall day of a SMART Survey
}
