% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_content.R
\name{f_find_codes}
\alias{f_find_codes}
\title{Find code to know the type of death if known}
\usage{
f_find_codes(smart_content)
}
\arguments{
\item{smart_content}{Content of smart survey}
}
\value{
[data.frame] 3 integers for three different types of death: unknow, injury and violence
}
\description{
Find code to know the type of death if known
}
