% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_one_survey.R
\name{f_generate_new_name}
\alias{f_generate_new_name}
\title{Function to extraction nutrition content of a SMART Survey}
\usage{
f_generate_new_name(log_file, country, infos)
}
\arguments{
\item{infos}{information extracted from the smart survey}
}
\value{
new_name of the smart survey
}
\description{
Function to extraction nutrition content of a SMART Survey
}
