% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admins_list.R
\name{f_location_possibility}
\alias{f_location_possibility}
\title{Function to return the right list of settlement depending of the country}
\usage{
f_location_possibility(country = "SOM")
}
\arguments{
\item{country}{Specified country of interest}
}
\value{
[data.frame] DataFrame with all the possibility of admin1 to match SMART Surveys infos
}
\description{
Function to return the right list of settlement depending of the country
}
\details{
Settlement lists (admin1)
}
