% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_one_survey.R
\name{f_save_smart}
\alias{f_save_smart}
\title{Function to save the SMART Survey in the right folder (the one with the new name or the Issue one)}
\usage{
f_save_smart(
  smart_folder,
  output_folder,
  file,
  new_name,
  infos,
  nutrition_data,
  mortality_data,
  clusters_data,
  metadata_data
)
}
\arguments{
\item{output_folder}{output folder to save the smart content}

\item{new_name}{new name of the smart survey used to generate a new folder}

\item{nutrition_data}{nutrition data containing in the smart}

\item{mortality_data}{mortality data containing in the smart}

\item{clusters_data}{clusters data containing in the smart}

\item{log_file}{log file}
}
\description{
Function to save the SMART Survey in the right folder (the one with the new name or the Issue one)
}
