% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_one_survey.R
\name{f_smart_is_issue}
\alias{f_smart_is_issue}
\title{Function to generate if there is an issue or not}
\usage{
f_smart_is_issue(infos)
}
\arguments{
\item{infos}{info of the smart surveys}
}
\description{
Function to generate if there is an issue or not
}
