% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_all_surveys.R
\name{f_sort_smart}
\alias{f_sort_smart}
\title{Function to read all the SMART Survey & save the different structure in the right folder & generate issue folder}
\usage{
f_sort_smart(smart_folder, output_folder, country, country_data)
}
\arguments{
\item{smart_folder}{Folder where the .as file are saved}

\item{output_folder}{Folder where the results need to be saved}
}
\description{
Function to read all the SMART Survey & save the different structure in the right folder & generate issue folder
}
