% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_one_survey.R
\name{f_update_log_file}
\alias{f_update_log_file}
\title{Function to extraction nutrition content of a SMART Survey}
\usage{
f_update_log_file(
  log_file,
  name_file,
  new_name,
  infos,
  nutrition_data,
  mortality_data,
  clusters_data,
  metadata_data
)
}
\arguments{
\item{log_file}{Log file}

\item{infos}{information extracted from the smart survey}

\item{nutrition_data}{nutrition data containing in the smart}

\item{mortality_data}{mortality data containing in the smart}

\item{clusters_data}{clusters data containing in the smart}
}
\value{
[data.frame] return the log file updated
}
\description{
Function to extraction nutrition content of a SMART Survey
}
