% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_infos.R
\name{find_date}
\alias{find_date}
\title{Find the date using .as information}
\usage{
find_date(smart_content)
}
\arguments{
\item{smart_content}{content of .as file}
}
\value{
Date - Day, Month, Year
}
\description{
Find the date using .as information
}
