% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_infos.R
\name{find_stratum_level}
\alias{find_stratum_level}
\title{Find the stratum level using .as information and settlement.csv file}
\usage{
find_stratum_level(name_file, smart_content, admin_list)
}
\arguments{
\item{name}{name of the .as file}
}
\value{
Two elements -- type of level and the name of the settlement/district/region/lhz
}
\description{
Find the stratum level using .as information and settlement.csv file
}
