% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_one_survey.R
\name{generate_log_file}
\alias{generate_log_file}
\title{Function to generate log file}
\usage{
generate_log_file(output_folder)
}
\value{
[data.frame] log file
}
\description{
Function to generate log file
}
