% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_infos.R
\name{is_individual_survey}
\alias{is_individual_survey}
\title{test if the survey is an individual or an aggregate one}
\usage{
is_individual_survey(smart_content)
}
\arguments{
\item{smart_content}{coontent of the SMART Survey}
}
\value{
TRUE is the .as file is an individual one
}
\description{
test if the survey is an individual or an aggregate one
}
